Refactor v1 to v2
