+++ components/actions.ts
@@ -2,44 +2,56 @@
 
 import { AnimeAP, getAnimes as getAnimesAP } from '@/lib/ap'
 import { getAnimeMAL, getUserAnimes, updateAnimeStatus } from '@/lib/mal'
-import { Anime } from './Scraper'
-import { cookies } from 'next/headers'
 import { getUser } from '@/lib/users'
+import { cookies } from 'next/headers'
+import { AnimeAPMAL, AnimeAPMALInfo } from './types'
 
-export async function getAnimes(username: string): Promise<Array<Anime>> {
+export async function getUserAnimesAP(
+    username: string,
+): Promise<Array<AnimeAP>> {
     const animes = await getAnimesAP(username)
+    return animes
+}
 
+export async function mapAPtoMAL(
+    apAnimes: Array<AnimeAP>,
+): Promise<Array<AnimeAPMAL>> {
     const animesData = await Promise.all(
-        animes.map(async (apAnime: AnimeAP) => {
+        apAnimes.map(async (apAnime: AnimeAP) => {
             const malAnime = await getAnimeMAL(apAnime)
 
             return {
                 ap: apAnime,
                 mal: malAnime,
+                info:
+                    malAnime === undefined
+                        ? AnimeAPMALInfo.NOT_FOUND
+                        : AnimeAPMALInfo.FOUND,
             }
         }),
     )
 
+    // Can be removed, just for debug
     const notfounds = animesData.filter((anime) => anime.mal === undefined)
-
     console.log('TOTAL: ' + animesData.length)
     console.log('NOT FOUND: ' + notfounds.length)
+    // Can be removed, just for debug
 
     return animesData
 }
 
-export async function getAnimesMal() {
+export async function getUserAnimesMAL() {
     const cookieStore = cookies()
     const userId = cookieStore.get('userId')?.value
 
     if (!userId) {
-        return undefined
+        throw new Error('Not logged in')
     }
 
     const user = await getUser(userId)
 
     if (!user) {
-        return undefined
+        throw new Error('User not found')
     }
 
     const animes = await getUserAnimes(user.access_token)
@@ -47,7 +59,7 @@ export async function getAnimesMal() {
     return animes
 }
 
-export async function updateAnimesMAL(animes: Array<Anime>) {
+export async function updateAnimesMAL(animes: Array<AnimeAPMAL>) {
     const cookieStore = cookies()
     const userId = cookieStore.get('userId')?.value
 
@@ -61,7 +73,29 @@ export async function updateAnimesMAL(animes: Array<Anime>) {
         return undefined
     }
 
-    const notUpdatedAnimes = await updateAnimeStatus(user.access_token, animes)
+    const notUpdatedAnimes = [] as Array<AnimeAPMAL>
+
+    const promises = animes.map(async (anime) => {
+        if (anime.mal === undefined) {
+            notUpdatedAnimes.push(anime)
+            return
+        }
+
+        const updatedAnime = await updateAnimeStatus(
+            user.access_token,
+            anime.mal.id,
+            anime.ap.status,
+            Number(anime.ap.eps_watched),
+        )
+
+        console.log('Hello: ' + updatedAnime)
+
+        if (!updatedAnime) {
+            notUpdatedAnimes.push(anime)
+        }
+    })
+
+    await Promise.all(promises)
 
     return notUpdatedAnimes
 }
